Rename BaseAuthenticator to AuthenticatorBase
